[@mantine/hooks] use-intersection: Fix docs and hook code
