--- docs/src/demos/hooks/use-intersection/index.ts
@@ -1,2 +1,0 @@
-export { UseIntersectionSimpleDemo } from './simple';
-export { UseIntersectionContainedDemo } from './contained';
