+++ src/mantine-hooks/src/use-intersection/use-intersection.ts
@@ -2,32 +2,31 @@ import { useCallback, useRef, useState } from 'react';
 
 export function useIntersection<T extends HTMLElement = any>(
   options?: ConstructorParameters<typeof IntersectionObserver>[1]
-): readonly [(element: T) => void, IntersectionObserverEntry | null] {
-  const [latestEntry, setLatestEntry] = useState<null | IntersectionObserverEntry>(null);
+) {
+  const [entry, setEntry] = useState<IntersectionObserverEntry>(null);
 
-  const ioRef = useRef<IntersectionObserver>();
+  const observer = useRef<IntersectionObserver>();
 
-  const refCallback = useCallback(
+  const ref = useCallback(
     (element: T | null) => {
-      if (ioRef.current) {
-        ioRef.current.disconnect();
-        ioRef.current = null;
+      if (observer.current) {
+        observer.current.disconnect();
+        observer.current = null;
       }
 
       if (element === null) {
-        setLatestEntry(null);
+        setEntry(null);
         return;
       }
 
-      ioRef.current = new IntersectionObserver(([entry]) => {
-        /* The hook supports observing just one element */
-        setLatestEntry(entry);
+      observer.current = new IntersectionObserver(([_entry]) => {
+        setEntry(_entry);
       }, options);
 
-      ioRef.current.observe(element);
+      observer.current.observe(element);
     },
     [options?.rootMargin, options?.root, options?.threshold]
   );
 
-  return [refCallback, latestEntry] as const;
+  return [ref, entry] as const;
 }
